+++ kernel/include/kernel/mem/kernel_vm.h
@@ -5,6 +5,6 @@
 
 extern void KERNEL_VMA();
 
-#define KERNEL_VM_START ((uint64_t) &KERNEL_VMA)
+#define KERNEL_VM_START ((uintptr_t) &KERNEL_VMA)
 
 #endif /* _KERNEL_MEM_KERNE_VM_H */
